Fix for the install script.

+++ scripts/install_craft.sh
@@ -8,14 +8,21 @@
 # @since     1.0.2
 # @license   MIT
 
-cd ..
+# Make sure the `.env.sh` exists
+if [[ ! -f ".env.sh" ]] ; then
+    echo 'File ".env.sh" is missing, aborting.'
+    exit
+fi
+
+source ".env.sh"
+
+cd $LOCAL_ROOT_PATH
 
 # Download and unzip the latest CraftCMS
 mkdir craft-install
 cd craft-install/
 wget -O latest.zip https://craftcms.com/latest.zip?accept_license=yes
 unzip latest.zip
-
 cd ..
 
 # Move folders into place
@@ -26,4 +33,5 @@ mv craft-install/craft/config/db.php craft/config/db.php
 # Clean up
 rm -rf craft-install
 
+# Set the appropriate permissions
 ./scripts/set_perms.sh
